+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - raise on `validate: true` in `CHECK` constraints https://github.com/plausible/ecto_ch/pull/124
+- add support for unsafe hints https://github.com/plausible/ecto_ch/pull/102
 
 ## 0.3.0 (2023-09-13)